requirements.txt
Pillow==9.5.0
pyenchant==3.2.2